move getNextProfileCompletionStep function to helper file.
