+++ src/utils/helpers.ts
@@ -4,6 +4,7 @@ import {
   InsertNotification,
   SelectChannel,
   SelectCommunity,
+  SelectProfile,
   SelectRole,
   SelectSpace,
   SelectUser,
@@ -447,4 +448,35 @@ export function GetSpaceURL(
 
   const URL = createAbsoluteUrl(path)
   return URL
+}
+
+export function getNextProfileCompletionStep(
+  profile?: SelectProfile | null
+): number {
+  if (!profile) {
+    return 1
+  }
+
+  if (profile.is_profile_completed === 1) {
+    return 4
+  }
+
+  const hasEducationData = Boolean(
+    profile.degree?.trim() ||
+      profile.institute?.trim() ||
+      profile.education_start_date?.trim() ||
+      profile.education_end_date?.trim()
+  )
+
+  if (hasEducationData) {
+    return 3
+  }
+
+  const hasBasicProfileData = Boolean(profile.bio?.trim())
+
+  if (hasBasicProfileData) {
+    return 2
+  }
+
+  return 1
 }
